Rename encoder mod to reply
